+++ api/src/opentrons/protocol_engine/types.py
@@ -14,7 +14,6 @@ class EngineStatus(str, Enum):
 
     IDLE = "idle"
     RUNNING = "running"
-    PAUSE_REQUESTED = "pause-requested"
     PAUSED = "paused"
     STOP_REQUESTED = "stop-requested"
     STOPPED = "stopped"