+++ app/src/organisms/RunTimeControl/index.tsx
@@ -14,14 +14,13 @@ import {
   Flex,
   Icon,
   IconName,
-  PrimaryBtn,
+  NewPrimaryBtn,
   Text,
   Tooltip,
   useHoverTooltip,
   ALIGN_CENTER,
   ALIGN_STRETCH,
   BORDER_RADIUS_1,
-  C_BLUE,
   DIRECTION_COLUMN,
   DISPLAY_FLEX,
   FONT_BODY_1_DARK_SEMIBOLD,
@@ -56,7 +55,55 @@ export function RunTimeControl(): JSX.Element | null {
   const pausedAt = useRunPauseTime()
   const completedAt = useRunCompleteTime()
 
-  const { play, pause, reset } = useRunControls()
+  const {
+    play,
+    pause,
+    reset,
+    isPlayRunActionLoading,
+    isPauseRunActionLoading,
+    isResetRunLoading,
+  } = useRunControls()
+
+  const [isRunActionLoading, setIsRunActionLoading] = React.useState(false)
+  const [lastRunAction, setLastRunAction] = React.useState<
+    'play' | 'pause' | 'reset' | null
+  >(null)
+
+  React.useEffect(() => {
+    if (isPlayRunActionLoading) {
+      setIsRunActionLoading(true)
+      setLastRunAction('play')
+    } else if (isPauseRunActionLoading) {
+      setIsRunActionLoading(true)
+      setLastRunAction('pause')
+    } else if (isResetRunLoading) {
+      setIsRunActionLoading(true)
+      setLastRunAction('reset')
+    }
+
+    if (isRunActionLoading) {
+      if (lastRunAction === 'play' && runStatus === RUN_STATUS_RUNNING) {
+        setIsRunActionLoading(false)
+      }
+      if (
+        lastRunAction === 'pause' &&
+        (runStatus === RUN_STATUS_PAUSED ||
+          runStatus === RUN_STATUS_PAUSE_REQUESTED)
+      ) {
+        setIsRunActionLoading(false)
+      }
+      if (lastRunAction === 'reset' && runStatus === RUN_STATUS_IDLE) {
+        setIsRunActionLoading(false)
+      }
+    }
+  }, [
+    isPlayRunActionLoading,
+    isPauseRunActionLoading,
+    isResetRunLoading,
+    isRunActionLoading,
+    lastRunAction,
+    runStatus,
+  ])
 
   let handleButtonClick = (): void => {}
   let buttonIconName: IconName | null = null
@@ -88,6 +135,11 @@ export function RunTimeControl(): JSX.Element | null {
     handleButtonClick = reset
   }
 
+  const buttonIcon =
+    buttonIconName != null ? (
+      <Icon name={buttonIconName} size={SIZE_1} marginRight={SPACING_2} />
+    ) : null
+
   return runStatus != null ? (
     <Flex flexDirection={DIRECTION_COLUMN} margin={SPACING_2}>
       <Text css={FONT_HEADER_DARK} marginBottom={SPACING_3}>
@@ -105,24 +157,25 @@ export function RunTimeControl(): JSX.Element | null {
           completedAt={completedAt}
         />
       ) : null}
-      <PrimaryBtn
+      <NewPrimaryBtn
         onClick={handleButtonClick}
         alignSelf={ALIGN_STRETCH}
-        backgroundColor={C_BLUE}
         borderRadius={BORDER_RADIUS_1}
         paddingTop={SPACING_2}
         paddingBottom={SPACING_2}
         justifyContent={JUSTIFY_CENTER}
         alignItems={ALIGN_CENTER}
         display={DISPLAY_FLEX}
-        disabled={disableRunCta}
+        disabled={disableRunCta || isRunActionLoading}
         {...targetProps}
       >
-        {buttonIconName != null ? (
-          <Icon name={buttonIconName} size={SIZE_1} marginRight={SPACING_2} />
-        ) : null}
+        {isRunActionLoading ? (
+          <Icon name="ot-spinner" size={SIZE_1} marginRight={SPACING_2} spin />
+        ) : (
+          buttonIcon
+        )}
         <Text fontSize={FONT_SIZE_DEFAULT}>{buttonText}</Text>
-      </PrimaryBtn>
+      </NewPrimaryBtn>
       {disableRunCta && (
         <Tooltip {...tooltipProps}>{t('run_cta_disabled')}</Tooltip>
       )}